Refactoring and fixed bugs
