--- src/common/styles/Container.module.css
@@ -1,12 +1,0 @@
-.container {
-    border: 1px solid;
-    height: 100%;
-    width: 80%;
-    margin: 0 auto;
-    display: flex;
-    align-items: flex-start;
-    padding-left: 50px;
-    justify-content: space-between;
-    flex-wrap: wrap;
-
-}
