+++ src/contacts/Сontacts.module.scss
@@ -7,8 +7,25 @@
   font-family: $fontFamily;
 
   .contactsContainer {
-    flex-direction: column;
-    align-items: center;
+    @include container();
+    @include flex(column, space-between, center, wrap);
+    input:-webkit-autofill,
+    input:-webkit-autofill:hover,
+    input:-webkit-autofill:focus
+    input:-webkit-autofill,
+    textarea:-webkit-autofill,
+    textarea:-webkit-autofill:hover
+    textarea:-webkit-autofill:focus,
+    select:-webkit-autofill,
+    select:-webkit-autofill:hover,
+    select:-webkit-autofill:focus {
+      border: 2px solid $borderColor;
+      -webkit-text-fill-color: $secondaryColor;
+      -webkit-box-shadow: 0 0 0 1000px transparent inset;
+      transition: background-color 5000s ease-in-out 0s;
+      background: -webkit-linear-gradient(top,  rgba(255,255,255,0) 0%,rgba(0,174,255,0.04) 50%,rgba(255,255,255,0) 51%,rgba(0,174,255,0.03) 100%);
+    }
+
   }
 
   .email {
@@ -59,7 +76,7 @@
       margin-bottom: 20px;
       outline: none;
       transition: all 0.3s ease;
-      height: 100px;
+      height: 150px;
       background: transparent;
       @include text(15px, 28px);
       color: $secondaryColor;
@@ -78,6 +95,7 @@
     transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
     letter-spacing: 0.2px;
     margin-bottom: 20px;
+    cursor: pointer;
 
     &:hover {
       border: 2px solid $primaryColor;
