+++ src/footer/Footer.module.scss
@@ -6,8 +6,8 @@
   font-family: $fontFamily;
 
   .footerContainer {
-    flex-direction: column;
-    align-items: center;
+    @include container();
+    @include flex(column, space-between, center, wrap);
 
     & h3 {
       @include text(16px, 28px);
@@ -21,6 +21,10 @@
         height: 45px;
         width: 45px;
         padding: 5px;
+        &:hover {
+          box-sizing: border-box;
+          transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1)
+        }
       }
     }
