+++ src/main/Main.module.scss
@@ -4,7 +4,7 @@
 .mainBlock {
   background-color: #121528;
   font-family: $fontFamily;
-  min-height: calc(100vh - 80);
+  min-height: calc(100vh - 80px);
   display: flex;
   align-items: center;
   //padding: 20px 0;
