+++ src/myProjects/myProject/MyProject.tsx
@@ -5,7 +5,8 @@ import Tilt from "react-parallax-tilt";
 type MyProjectType = {
     title: string
     description: string
-    style?: any
+    style: any
+    link: string
 }
 export const MyProject = (props: MyProjectType) => {
     return (
@@ -13,7 +14,7 @@ export const MyProject = (props: MyProjectType) => {
         <Tilt tiltMaxAngleX={10} tiltMaxAngleY={10}>
             <div className={style.portfolio}>
                 <div className={style.projectImgContainer} style={props.style}>
-                    <a className={style.myProjectButton} href={'ya.ru'}>More</a>
+                    <a className={style.myProjectButton} href={props.link}>More</a>
                 </div>
                 <div className={style.content}>
                     <h3 className={style.projectTitle}>{props.title}</h3>
