+++ src/skills/Skills.tsx
@@ -1,5 +1,5 @@
 import style from './Skills.module.css'
-import styleContainer from './../common/styles/Container.module.css'
+import styleContainer from '../common/styles/Container1.module.css'
 import {Skill} from "./skill/Skill";
 
 export const Skills = () => {
